fix: 修正剧情 VM 世界重载地图参数

+++ include/openswd3/world_map/legacy_maps_world_database.hpp
@@ -33,7 +33,7 @@ struct LegacyMapsWorldHeader {
 };
 
 struct LegacyWorldLoadRequest {
-    compat::u16 logical_map_id{};
+    compat::u32 logical_map_id{};
     compat::u32 tile_x{};
     compat::u32 tile_y{};
     compat::u16 action_id{};
